server/requirements.txt
websockets==8.0.2
attrs
pickledb
numpy
